fix loginForm is not prefilled with address
When the loginForm promise is resolved, the url
arg loginWith is read and stored in the mailAddress
Stream. The loginForm promise is only resolved
in the oncreate function of the loginForm.
Since the oncreate function is not called again
when switching from saved accounts to create a new
account, the promise won't be resolved again.

Moving the promise.resolve() to the onupdate
function solves the issue.

fix #5810