fix missing object store causing uncaught error reports
our indexedDb implementation is not very clear and the original
error points to some timing or error handling problem in the
initialization or migration sequence.
this patch stops these errors from becoming user-visible by ignoring
them.
if the indexDB is available after all, the user can still add the rule
later, if it is not, indexing will likely be disabled at some later
point.

#5915